Converters e Ajustes em Impl e DAO

+++ FinancasPessoais/src/main/java/com/financaspessoais/service/TransferenciaService.java
@@ -11,5 +11,5 @@ public interface TransferenciaService {
 
 	List<Transferencia> listarPorUsuario();
 
-	public Transferencia buscar(Integer id);
+	public Transferencia buscar(Long id);
 }
